+++ reikna/core/transformation.py
@@ -4,7 +4,7 @@ import numpy
 
 import reikna.cluda.dtypes as dtypes
 from reikna.cluda import Module
-from reikna.helpers import AttrDict, product, wrap_in_tuple, template_func, template_for
+from reikna.helpers import AttrDict, product, wrap_in_tuple, template_def, template_for
 
 
 TEMPLATE = template_for(__file__)
@@ -153,7 +153,7 @@ class Transformation:
         if snippet is None:
             snippet = "${" + self.outputs[0] + ".store}(${" + self.inputs[0] + ".load});"
 
-        self.snippet_template = template_func(
+        self.snippet_template = template_def(
             self.outputs + self.inputs + self.scalars,
             snippet)